+++ card/src/main/java/com/banksalad/collectmydata/card/common/db/entity/LoanLongTermEntity.java
@@ -40,7 +40,7 @@ public class LoanLongTermEntity extends BaseEntity {
   private String organizationId;
 
   @Column(nullable = false)
-  private Integer loanLongTermNo;
+  private Short loanLongTermNo;
 
   @Column(nullable = false)
   private String loanDtime;
